requirements.txt
pyannote.audio==3.1.1
torch==2.1.0
torchaudio==2.1.0
transformers==4.36.0
huggingface_hub==0.20.0
soundfile==0.12.1
numpy==1.24.3
